split spec file into 2 files
